requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
ag2[openai]
python-dotenv>=1.0.0
